+++ packages/wonder-blocks-breadcrumbs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @khanacademy/wonder-blocks-breadcrumbs
 
+## 2.1.3
+
+### Patch Changes
+
+-   Updated dependencies [674a1e5c]
+-   Updated dependencies [674a1e5c]
+    -   @khanacademy/wonder-blocks-core@6.0.0
+
+## 2.1.2
+
+### Patch Changes
+
+-   Updated dependencies [1344436f]
+    -   @khanacademy/wonder-blocks-core@5.4.0
+
 ## 2.1.1
 
 ### Patch Changes