+++ packages/wonder-blocks-dropdown/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @khanacademy/wonder-blocks-dropdown
 
+## 3.0.21
+
+### Patch Changes
+
+-   Updated dependencies [8c77f29d]
+-   Updated dependencies [674a1e5c]
+-   Updated dependencies [674a1e5c]
+-   Updated dependencies [674a1e5c]
+    -   @khanacademy/wonder-blocks-clickable@4.0.0
+    -   @khanacademy/wonder-blocks-modal@4.0.19
+    -   @khanacademy/wonder-blocks-core@6.0.0
+    -   @khanacademy/wonder-blocks-icon@2.1.1
+    -   @khanacademy/wonder-blocks-layout@2.0.17
+    -   @khanacademy/wonder-blocks-search-field@2.1.7
+    -   @khanacademy/wonder-blocks-typography@2.1.3
+
+## 3.0.20
+
+### Patch Changes
+
+-   Updated dependencies [ec6a33a4]
+-   Updated dependencies [1344436f]
+    -   @khanacademy/wonder-blocks-icon@2.1.0
+    -   @khanacademy/wonder-blocks-core@5.4.0
+    -   @khanacademy/wonder-blocks-search-field@2.1.6
+    -   @khanacademy/wonder-blocks-modal@4.0.18
+    -   @khanacademy/wonder-blocks-clickable@3.1.3
+    -   @khanacademy/wonder-blocks-layout@2.0.16
+    -   @khanacademy/wonder-blocks-typography@2.1.2
+
 ## 3.0.19
 
 ### Patch Changes