+++ packages/wonder-blocks-progress-spinner/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @khanacademy/wonder-blocks-progress-spinner
 
+## 2.0.17
+
+### Patch Changes
+
+-   Updated dependencies [674a1e5c]
+-   Updated dependencies [674a1e5c]
+    -   @khanacademy/wonder-blocks-core@6.0.0
+
+## 2.0.16
+
+### Patch Changes
+
+-   Updated dependencies [1344436f]
+    -   @khanacademy/wonder-blocks-core@5.4.0
+
 ## 2.0.15
 
 ### Patch Changes