+++ packages/wonder-blocks-toolbar/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @khanacademy/wonder-blocks-toolbar
 
+## 3.0.17
+
+### Patch Changes
+
+-   Updated dependencies [674a1e5c]
+-   Updated dependencies [674a1e5c]
+    -   @khanacademy/wonder-blocks-core@6.0.0
+    -   @khanacademy/wonder-blocks-typography@2.1.3
+
+## 3.0.16
+
+### Patch Changes
+
+-   Updated dependencies [1344436f]
+    -   @khanacademy/wonder-blocks-core@5.4.0
+    -   @khanacademy/wonder-blocks-typography@2.1.2
+
 ## 3.0.15
 
 ### Patch Changes